video-generators/ChessFlix/pgn-etl/requirements.txt
CairoSVG==2.4.2
moviepy==1.0.0 #  1.0.1 introduced bugs
requests==2.23.0
redis==3.4.1
imageio==2.8.0
python-chess==0.30.1


# To get this to run on windows, you need to follow this
# https://weasyprint.readthedocs.io/en/stable/install.html#gtk64installer
# then restart computer